% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{apply.labels}
\alias{apply.labels}
\title{Apply Hmisc-brand labels to all variables in a dataframe or tibble. Deprecated in favor of apply_lables.}
\usage{
apply.labels(df, labels)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{labels}{A vector of quoted label of length ncol(df).}
}
\value{
A data frame or tibble.
}
\description{
Apply Hmisc-brand labels to all variables in a dataframe or tibble. Deprecated in favor of apply_lables.
}
\examples{
foo <- tibble(a = c(1,2,3),
              b = c(4, 5,6)) \%>\%
  apply_labels(c("A", "B"))
label(foo)
}
