% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trahelyk.R
\name{apply_label}
\alias{apply_label}
\title{Apply an Hmisc-brand label to a single variable in a dataframe or tibble}
\usage{
apply_label(df, x, lbl)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{x}{Unquoted name of the column to be labeled.}

\item{lbl}{Quoted label.}
}
\value{
A data frame or tibble.
}
\description{
Apply an Hmisc-brand label to a single variable in a dataframe or tibble
}
\examples{
foo <- tibble(a = c(1,2,3),
              b = c(4, 5,6)) \%>\%
  apply_label(a, "A") \%>\%
  apply_label(b, "B")
  
label(foo)
}
