% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trahelyk.R
\name{apply_label_vector}
\alias{apply_label_vector}
\title{Apply a character vector as labels for a data frame}
\usage{
apply_label_vector(df, labels)
}
\arguments{
\item{df}{}

\item{labels}{}
}
\value{

}
\description{
Apply a character vector as labels for a data frame
}
