% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trahelyk.R
\name{apply_labels}
\alias{apply_labels}
\title{Apply Hmisc-brand labels to one or more variables in a dataframe or tibble}
\usage{
apply_labels(df, ...)
}
\arguments{
\item{df}{A data frame or tibble.}

\item{...}{Name-value pairs of column names and labels}
}
\value{
A data frame or tibble.
}
\description{
Apply Hmisc-brand labels to one or more variables in a dataframe or tibble
}
\examples{
foo <- tibble(a = c(1,2,3),
              b = c(4, 5,6)) \%>\%
  apply_labels(a = "A", 
               b = "B"))
label(foo)
}
