% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trahelyk.R
\name{ci_fmt}
\alias{ci_fmt}
\title{Calculate a 95% CI and return it as a formatted string}
\usage{
ci_fmt(x, d = 2, dist = "normal")
}
\arguments{
\item{x}{A numeric or integer vector}

\item{d}{An integer specifying the number of significant digits to be presented in the point estimate and interval.}

\item{dist}{A character string specifying the distribution from which the interval should be calculated. Currently supports "normal" or "t".}
}
\value{
A character string with the point estimate of the mean, followed by the interval in parentheses.
}
\description{
Calculate a 95% CI and return it as a formatted string
}
