% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trahelyk.R
\name{clear_labels}
\alias{clear_labels}
\title{Remove all Hmisc labels from a data frame}
\usage{
clear_labels(x)
}
\arguments{
\item{x}{Data frame to clean}
}
\value{
Data frame
}
\description{
Remove all Hmisc labels from a data frame
}
