% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trahelyk.R
\name{cnt}
\alias{cnt}
\title{Count observations and number of unique identifiers}
\usage{
cnt(DT = 0, current_count = NULL, guid = c(getOption("guid")))
}
\arguments{
\item{DT}{A data.table, tibble, or data.frame}

\item{current_count}{Previous list created by cnt(), containing the last count}

\item{guid}{Name of the globally unique identifier to count}
}
\value{
List object containing the number of observations in element one, and the number of unique guids in element two
}
\description{
Count observations and number of unique identifiers
}
