% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trahelyk.R
\name{fact0rny}
\alias{fact0rny}
\title{Return a 2-level factor with labels "No" and "Yes"}
\usage{
fact0rny(x)
}
\arguments{
\item{x}{A binary (0,1) numeric vector}
}
\value{
A factor with levels and lables c("No", "Yes")
}
\description{
Return a 2-level factor with labels "No" and "Yes"
}
