% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_tableone.R
\name{ft.tbl1}
\alias{ft.tbl1}
\title{Present a tableone table as a flextable}
\usage{
ft.tbl1(x, col_keys = "", include_combined = FALSE)
}
\arguments{
\item{x}{A tbl1 object}

\item{col_keys}{columns names/keys to display (passed on to flextable())}

\item{include_combined}{logical indicating whether the table should include the combined column.}
}
\value{
A flextable object
}
\description{
Present a tableone table as a flextable
}
