% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_tableone.R
\name{ft.tbl1w}
\alias{ft.tbl1w}
\title{Present a one-way table as a flextable}
\usage{
ft.tbl1w(x, col_keys = "")
}
\arguments{
\item{x}{A tbl1w object}

\item{col_keys}{columns names/keys to display (passed on to flextable())}
}
\value{
A flextable object
}
\description{
Present a one-way table as a flextable
}
