% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trahelyk.R
\name{is_unique}
\alias{is_unique}
\title{Determine whether a column in a data frame is unique across all rows}
\usage{
is_unique(x, guid, quiet = FALSE)
}
\arguments{
\item{x}{A data frame}

\item{guid}{Name of the column to evaluate for uniqueness}

\item{quiet}{Logical. If FALSE, the function echoes the number of unique values of `guid`; if TRUE, it only returns a logical.}
}
\value{
Logical indicating whether the guid is unique across rows
}
\description{
Determine whether a column in a data frame is unique across all rows
}
\examples{
foo <- tibble(A = c(1, 2, 3, 4, 5),
              B = c("A", "B", "B", "C", "D"))
is_unique(foo, "A")
is_unique(foo, "B")
}
