% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trahelyk.R
\name{look}
\alias{look}
\title{Display a summary and SD of a numeric vector}
\usage{
look(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Summmary and SD
}
\description{
Display a summary and SD of a numeric vector
}
\examples{
foo <- rnorm(n=20)
look(foo)
}
