% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.R
\name{md.tableone}
\alias{md.tableone}
\title{Display a tableone table in RMarkdown}
\usage{
\method{md}{tableone}(x, word = FALSE, wordstyles = NA)
}
\arguments{
\item{word}{Logical indicating whether the output should be formatted for an MS Word}

\item{wordstyles}{Name of the MS Word stylesheet to be used for formatting. NA if word==FALSE.}

\item{A}{tableone object, produced by make_tableone}
}
\value{
Side effects.
}
\description{
Display a tableone table in RMarkdown
}
