% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot a ROC curve based on a roc_obj object}
\usage{
plot_roc(roc_obj, thresh_pos = "top", thresh_size = 7)
}
\arguments{
\item{roc_obj}{Object generated by roc_obj()}

\item{thresh_pos}{Character vector specifying whether the plot should place Youden's criterion above or below its point on the curve. Valid options include "top" or "bottom". Defaults to "top."}

\item{thresh_size}{Numeric specifying the font size for Youden's criterion. Defaults to 7.}
}
\value{
A ggplot object
}
\description{
Plot a ROC curve based on a roc_obj object
}
