% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_models.R
\name{present_mdl.glm}
\alias{present_mdl.glm}
\title{Format and present a LaTeX table that summarizes a glm model object}
\usage{
\method{present_mdl}{glm}(mdl, d = 3, intercept = FALSE, varnames = NULL,
  lbl = "", caption = "", link = exp)
}
\arguments{
\item{mdl}{A glm object produced by stats::glm()}

\item{d}{An integer specifying the number of significant digits to be presented in the point estimate and interval. Defaults to 3.}

\item{intercept}{Logical indicating whether the intercept should be included in the presentation of the model. Defaults to FALSE.}

\item{varnames}{Character vector containing the English-version names of the independent variables in themodel. Defaults to names(mdl$coefficients)}

\item{lbl}{LaTeX label to attach to the presented table.}

\item{caption}{Caption to print above the table.}

\item{link}{Function to be applied to beta coefficients to yield, for example, odds ratios. I know this is not the link function. Defaults to exp.}
}
\value{
A formatted LaTeX table containing estimates of model parameters and some simple fit statistics.
}
\description{
Format and present a LaTeX table that summarizes a glm model object
}
