% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_models.R
\name{present_mdl}
\alias{present_mdl}
\title{Generic function to format and present a LaTeX table that summarizes a model object.}
\usage{
present_mdl(
  mdl,
  coef.head = "",
  link = identity,
  footer = c(),
  d = 3,
  intercept = TRUE,
  varnames,
  lbl = "",
  caption = "",
  corPars = FALSE
)
}
\arguments{
\item{mdl}{A model object}

\item{link}{Function to be applied to beta coefficients to yield, for example, odds ratios. I know this is not the link function. Defaults to identity.}

\item{footer}{Text to add to the table footer, printed below the formatted LaTeX table.}

\item{d}{An integer specifying the number of significant digits to be presented in the point estimate and interval. Defaults to 3.}

\item{intercept}{Logical indicating whether the intercept should be included in the presentation of the model. Defaults to TRUE.}

\item{varnames}{Character vector containing the English-version names of the independent variables in themodel. Defaults to names(mdl$coefficients)}

\item{lbl}{LaTeX label to attach to the presented table.}

\item{caption}{Caption to print above the table.}

\item{corPars}{Logical indicating whether correlation parameters should be included in the presentation of the model (for GEE models). Defaults to FALSE.}

\item{coef_head}{Header text to place at the top of the coefficients column in the formatted table.}
}
\value{
A formatted LaTeX table containing estimates of model parameters and some simple fit statistics.
}
\description{
Generic function to format and present a LaTeX table that summarizes a model object.
}
