% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc_obj}
\alias{roc_obj}
\title{Generate roc_obj object that is handy for summarizing discrimination statistics and for passing
to other functions for presentation}
\usage{
roc_obj(x, y, xlab = "", units = "", ci = FALSE, rnd_thresh = 0,
  rnd_discr = 2)
}
\arguments{
\item{x}{Numeric or ordinal vector of predictors}

\item{y}{Factor, numeric, or character vector of responses}

\item{units}{Character specifying units for formatting Youden's threshold}

\item{ci}{Logical indicating whether bootstrapped confidence intervals should be calculated}

\item{rnd_thresh}{Rounding precision (in decimal places) for Youden's threshold (default is 0)}

\item{rnd_discr}{Rounding precision for discrimination statistics (sensitivity and specificity; default is 2)}
}
\value{
A roc_obj object
}
\description{
Generate roc_obj object that is handy for summarizing discrimination statistics and for passing
to other functions for presentation
}
