% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trahelyk.R
\name{sem}
\alias{sem}
\title{Calculate a standard error}
\usage{
sem(x = NA, s = NA, n = NA)
}
\arguments{
\item{x}{A numeric vector}

\item{s}{A standard deviation}

\item{n}{N corresponding to s}
}
\value{
Standard error of the mean (numeric)
}
\description{
Calculate a standard error
}
