% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.table.one.R
\name{tableoneway}
\alias{tableoneway}
\title{Produces a tbl1w object, which contains summary statistics without grouping or bivariate tests.}
\usage{
tableoneway(
  df,
  summaryTypes = NULL,
  d = 1,
  p.digits = 3,
  fisher.simulate.p = FALSE,
  trunc_binary = TRUE,
  lbl = "",
  caption = ""
)
}
\arguments{
\item{df}{A data frame or tibble with Hmisc-brand labels.}

\item{d}{Number of significant digits to report in (non-p-value) numbers in the tableone object.}

\item{p.digits}{Number of significant digits to report in p-values.}

\item{fisher.simulate.p}{If TRUE use simulated p-values for Fisher's exact test. Default is FALSE.}

\item{trunc_binary}{If TRUE print only one row for categorical variables with 2 levels. If FALSE, binary variables are printed on two rows. Default is TRUE.}

\item{lbl}{LaTeX label to be passed to tx() for labeling the table in a LaTeX document.}

\item{caption}{Text to be passed to tx() or md() to caption the table.}

\item{testTypes}{A vector of length nrow(df) containing "t" for a t-test or "w" for a Wilcoxon rank-sum test; applies only to continuous variables. Leave NULL to allow the function to choose for you based on the distance between the mean and median.}
}
\value{
A tbl1w object that can be formatted by tx() or md().
}
\description{
Produces a tbl1w object, which contains summary statistics without grouping or bivariate tests.
}
