% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trahelyk.R
\name{tocase}
\alias{tocase}
\title{Change case of a character vector}
\usage{
tocase(x, case)
}
\arguments{
\item{x}{a character vector}

\item{case}{one of "upper", "lower", "title", or "sentence".}
}
\value{
A character vector of the same length, with the first letter of each word, separated by a space, capitalized.
}
\description{
Change case of a character vector
}
\examples{
x <- "the quick BROWN fox jumps over the LaZy Dog."
totitle(x)
}
