% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.table.one.R
\name{trahbulate}
\alias{trahbulate}
\title{Tabulate a set of variables by an N-dimensional grouping variable.}
\usage{
trahbulate(
  df,
  grpvar,
  testTypes = NULL,
  d = 1,
  p.digits = 3,
  fisher.simulate.p = FALSE,
  trunc_binary = TRUE,
  lbl = "",
  caption = ""
)
}
\arguments{
\item{df}{A data frame or tibble with Hmisc-brand labels.}

\item{grpvar}{The quoted name of a grouping variable that splits df into N cohorts.}

\item{testTypes}{A vector of length nrow(df) containing "mean" for means and SDs or "median" medians and IQRs; applies only to continuous variables. Leave NULL to allow the function to choose for you based on the distance between the mean and median.}

\item{d}{Number of significant digits to report in numbers in the tableone object.}

\item{p.digits}{Not used.}

\item{fisher.simulate.p}{Not used.}

\item{trunc_binary}{If TRUE print only one row for categorical variables with 2 levels. If FALSE, binary variables are printed on two rows, with the variable name on a row above. Default is TRUE.}

\item{lbl}{LaTeX label to be passed to tx() for labeling the table in a LaTeX document.}

\item{caption}{Text to be passed to tx() or md() to caption the table.}
}
\value{
A tableone object that can be formatted by tx() or md().
}
\description{
Tabulate a set of variables by an N-dimensional grouping variable.
}
