% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{tx.lme}
\alias{tx.lme}
\title{Present an nlme-produced lme model object in a LaTeX table}
\usage{
\method{tx}{lme}(
  x,
  d = 3,
  intercept = TRUE,
  varnames = NULL,
  lbl = "",
  caption = "",
  coef.head = "Coefficient",
  link = identity,
  print_mdl = TRUE
)
}
\arguments{
\item{d}{An integer specifying the number of significant digits to be presented in the table.}

\item{intercept}{Logical indicating whether the table should include a row for the fixed intercept. Default is TRUE.}

\item{varnames}{Character vector of length length(names(mdl$coefficients$fixed)) specifying the names of fixed effects in the model.}

\item{lbl}{LaTeX label for the table.}

\item{caption}{Caption for the table.}

\item{coef.head}{Word or phrase to use as a header for coefficient estimates.}

\item{link}{Link function. Default is identity().}

\item{print_mdl}{Logical indicating whether the summary table should be wrapped in LaTeX.}

\item{mdl}{An lme object}
}
\value{
A LaTeX or clear-text table.
}
\description{
Present an nlme-produced lme model object in a LaTeX table
}
