% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{tx.tableone_stratified}
\alias{tx.tableone_stratified}
\title{Present two tableone objects in a stratified LaTeX table}
\usage{
\method{tx}{tableone_stratified}(t1, t2, head1, head2, size = "footnotesize")
}
\arguments{
\item{t1}{A tableone object}

\item{t2}{A second tableone object}

\item{head1}{Header for the first stratum}

\item{head2}{Header for the second stratum}

\item{size}{Font size}
}
\value{
A LaTeX table.
}
\description{
Present two tableone objects in a stratified LaTeX table
}
