% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{tx.tableoneway}
\alias{tx.tableoneway}
\title{Present a tableoneway object in a LaTeX table}
\usage{
\method{tx}{tableoneway}(x, size = "footnotesize")
}
\arguments{
\item{x}{A tableoneway object}

\item{size}{Font size}
}
\value{
A LaTeX or clear-text table.
}
\description{
Present a tableoneway object in a LaTeX table
}
