% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{tx.trahble}
\alias{tx.trahble}
\title{Display a trahble-based tabulation in LaTeX}
\usage{
\method{tx}{trahble}(x, size = "footnotesize", d = 3, include_n = FALSE)
}
\arguments{
\item{x}{A trahble object, produced by trahbulate.}

\item{size}{Font size}

\item{d}{Number of digits to the right of the decimal place for numerics}

\item{include_n}{Logical indicating whether the tabulation should include a column showing the number of non-missing observations for each variable}
}
\description{
Display a trahble-based tabulation in LaTeX
}
