% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_models.R
\name{tx_mdl.glm}
\alias{tx_mdl.glm}
\title{Present a glm-produced model object in a LaTeX table}
\usage{
\method{tx_mdl}{glm}(mdl, d = 3, intercept = FALSE, varnames = NULL,
  lbl = "", caption = "", coef.head = "OR", FUN = exp, footer = c(),
  print_mdl = TRUE)
}
\arguments{
\item{mdl}{A glm object}

\item{d}{An integer specifying the number of significant digits to be presented in the table.}

\item{intercept}{Logical indicating whether the table should include a row for the fixed intercept. Default is FALSE.}

\item{varnames}{Character vector of length length(names(mdl$coefficients)) specifying the names of terms in the model.}

\item{lbl}{LaTeX label for the table.}

\item{caption}{Caption for the table.}

\item{coef.head}{Word or phrase to use as a header for coefficient estimates. Default is "OR".}

\item{FUN}{Function to apply to coefficients before presenting. Default is exp().}

\item{print_mdl}{Logical indicating whether the summary table should be wrapped in LaTeX.}
}
\value{
A LaTeX or clear-text table.
}
\description{
Present a glm-produced model object in a LaTeX table
}
