% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_models.R
\name{tx_mdl.lm}
\alias{tx_mdl.lm}
\title{Present a lm-produced model object in a LaTeX table}
\usage{
\method{tx_mdl}{lm}(
  mdl,
  d = 3,
  intercept = TRUE,
  varnames = NULL,
  lbl = "",
  caption = "",
  coef.head = "Coefficient",
  footer = c(),
  FUN = identity,
  print_mdl = TRUE
)
}
\arguments{
\item{mdl}{A lm object}

\item{d}{An integer specifying the number of significant digits to be presented in the table.}

\item{intercept}{Logical indicating whether the table should include a row for the fixed intercept. Default is TRUE.}

\item{varnames}{Character vector of length length(names(mdl$coefficients)) specifying the names of terms in the model.}

\item{lbl}{LaTeX label for the table.}

\item{caption}{Caption for the table.}

\item{coef.head}{Word or phrase to use as a header for coefficient estimates. Default is "OR".}

\item{FUN}{Function to apply to coefficients before presenting. Default is identity().}

\item{print_mdl}{Logical indicating whether the summary table should be wrapped in LaTeX. Default is TRUE.}
}
\value{
A LaTeX or clear-text table.
}
\description{
Present a lm-produced model object in a LaTeX table
}
