% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.R
\name{md.tblnw}
\alias{md.tblnw}
\title{Display a trahble-based tabulation in RMarkdown}
\usage{
\method{md}{tblnw}(x, word = FALSE, wordstyles = NA)
}
\arguments{
\item{word}{Logical indicating whether the output should be formatted for an MS Word .}

\item{wordstyles}{Name of the MS Word stylesheet to be used for formatting. NA if word==FALSE.}

\item{A}{trahble object, produced by trahbulate.}
}
\description{
Display a trahble-based tabulation in RMarkdown
}
