% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{tx.tableoneway}
\alias{tx.tableoneway}
\title{Present a tableoneway object in a LaTeX table}
\usage{
\method{tx}{tableoneway}(x, size = "footnotesize", include_n = FALSE)
}
\arguments{
\item{x}{A tableoneway object}

\item{size}{Font size}

\item{include_n}{Include a column showing non-missing count for each variable? Logical; default=FALSE.}
}
\value{
A LaTeX or clear-text table.
}
\description{
Present a tableoneway object in a LaTeX table
}
